requirements.txt
langchain
langchain-openai
promptfoo
sentence-transformers
numpy
requests
python-dotenv
langchain-community
pypdf
